+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/ProgressiveMediaSource.java
@@ -96,6 +96,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
     private int singleTrackId;
     @Nullable private Format singleTrackFormat;
     private boolean loadOnlySelectedTracks;
+    private boolean experimentalEnableHagcPlayback;
 
     /**
      * Creates a new factory for {@link ProgressiveMediaSource}s.
@@ -186,6 +187,19 @@ public final class ProgressiveMediaSource extends BaseMediaSource
       this.drmSessionManagerProvider = drmSessionManagerProvider;
       this.loadErrorHandlingPolicy = loadErrorHandlingPolicy;
       this.continueLoadingCheckIntervalBytes = continueLoadingCheckIntervalBytes;
+      this.experimentalEnableHagcPlayback = true;
+    }
+
+    /**
+     * Sets whether to enable experimental HAGC (ST 2094-50) metadata playback support.
+     *
+     * @param enableHagcPlayback Whether experimental HAGC metadata playback is enabled.
+     * @return This factory, for convenience.
+     */
+    @CanIgnoreReturnValue
+    public Factory setExperimentalEnableHagcPlayback(boolean enableHagcPlayback) {
+      this.experimentalEnableHagcPlayback = enableHagcPlayback;
+      return this;
     }
 
     @CanIgnoreReturnValue
@@ -303,6 +317,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
           loadErrorHandlingPolicy,
           continueLoadingCheckIntervalBytes,
           loadOnlySelectedTracks,
+          experimentalEnableHagcPlayback,
           singleTrackId,
           singleTrackFormat,
           downloadExecutorSupplier);
@@ -326,6 +341,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
   private final LoadErrorHandlingPolicy loadableLoadErrorHandlingPolicy;
   private final int continueLoadingCheckIntervalBytes;
   private final boolean loadOnlySelectedTracks;
+  private final boolean experimentalEnableHagcPlayback;
 
   /**
    * The ID passed to {@link Factory#enableLazyLoadingWithSingleTrack(int, Format)}. Only valid if
@@ -361,6 +377,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
       LoadErrorHandlingPolicy loadableLoadErrorHandlingPolicy,
       int continueLoadingCheckIntervalBytes,
       boolean loadOnlySelectedTracks,
+      boolean experimentalEnableHagcPlayback,
       int singleTrackId,
       @Nullable Format singleTrackFormat,
       @Nullable Supplier<ReleasableExecutor> downloadExecutorSupplier) {
@@ -371,6 +388,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
     this.loadableLoadErrorHandlingPolicy = loadableLoadErrorHandlingPolicy;
     this.continueLoadingCheckIntervalBytes = continueLoadingCheckIntervalBytes;
     this.loadOnlySelectedTracks = loadOnlySelectedTracks;
+    this.experimentalEnableHagcPlayback = experimentalEnableHagcPlayback;
     this.singleTrackFormat = singleTrackFormat;
     this.singleTrackId = singleTrackId;
     this.timelineIsPlaceholder = true;
@@ -432,6 +450,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
         localConfiguration.customCacheKey,
         continueLoadingCheckIntervalBytes,
         loadOnlySelectedTracks,
+        experimentalEnableHagcPlayback,
         singleTrackId,
         singleTrackFormat,
         Util.msToUs(localConfiguration.imageDurationMs),